refactor(core): move hw entropy into separate file
