+++ core/embed/unix/common.c
@@ -23,7 +23,6 @@
 
 #include "common.h"
 #include "display.h"
-#include "memzero.h"
 
 extern void main_clean_exit();
 
@@ -108,7 +107,3 @@ error_shutdown(const char *line1, const char *line2, const char *line3,
 void hal_delay(uint32_t ms) { usleep(1000 * ms); }
 
 void wait_random(void) {}
-
-uint8_t HW_ENTROPY_DATA[HW_ENTROPY_LEN];
-
-void collect_hw_entropy(void) { memzero(HW_ENTROPY_DATA, HW_ENTROPY_LEN); }
