+++ core/embed/unix/common.h
@@ -58,8 +58,4 @@ error_shutdown(const char *line1, const char *line2, const char *line3,
 void hal_delay(uint32_t ms);
 void wait_random(void);
 
-void collect_hw_entropy(void);
-#define HW_ENTROPY_LEN (12 + 32)
-extern uint8_t HW_ENTROPY_DATA[HW_ENTROPY_LEN];
-
 #endif
