+++ core/embed/unix/entropy.h
@@ -1,0 +1,28 @@
+/*
+ * This file is part of the Trezor project, https://trezor.io/
+ *
+ * Copyright (c) SatoshiLabs
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __UNIX_ENTROPY_H__
+#define __UNIX_ENTROPY_H__
+
+#include <stdint.h>
+
+#define HW_ENTROPY_LEN (12 + 32)
+void collect_hw_entropy(void);
+extern uint8_t HW_ENTROPY_DATA[HW_ENTROPY_LEN];
+#endif
